move: "登録者入室時バンする" 機能関連のコードを 該当するファイルに移動
関連するHarmonyPatchを分離して``HarmonyPatch``に移動した。
(移動前の)``HaemonyPatch``に直接記載していた処理を関数として分離し、``ImmigrationCheck.CS``と``FriendCodeImmigrationPatch.cs``に移動した。
